MET-1111 Changed the data type of sigma to double
Former-commit-id: 1ff7f92da65b0ec054b0c67eef7b3db899ae7a95

+++ src/basic/vx_config/config_constants.h
@@ -244,7 +244,7 @@ struct InterpInfo {
    int         n_interp;   // Number of interpolation types
    StringArray method;     // Interpolation methods
    IntArray    width;      // Interpolation widths
-   int         sigma;      // Sigma for Gaussian
+   double      sigma;      // Sigma for Gaussian
    GridTemplateFactory::GridTemplates shape; // Interpolation shape
 
    void        clear();
@@ -266,7 +266,7 @@ struct RegridInfo {
                             // or explicit grid definition.
    InterpMthd   method;     // Regridding method
    int          width;      // Regridding width
-   int          sigma;      // Sigma for Gaussian
+   double       sigma;      // Sigma for Gaussian
    GridTemplateFactory::GridTemplates shape; // Interpolation shape
    RegridInfo();
